requirements.txt
pandas
scikit-learn
pymysql
mysql-connector-python
sqlalchemy